# Minimize updated areas of editors for playhead movement