While playing, the song editor and pattern editors will update
periodically to show the progress of the playhead across the song
or pattern. The backgrounds are cached as pixmaps so this is a lot
faster than redrawing the whole song or pattern, but blitting the pixmap can
still take significant amounts of time when large areas of screen space
are taken up by these widgets (particularly on, say, a Raspberry
Pi attached to a 2560x1440 display). It's unclear if Qt is using
any hardware acceleration for this, but I've seen large amounts of
CPU time used in Linux and macOS profiling.

This change modifies the timer-driven updates to call update() on
only the minimal area necessary to correctly keep the widget fresh,
and erase the old playhead and render the new one.

There's an annoying amount of code duplication here, as there's 
no common ancestor class to fold the song editor geometry into.
That feels ripe for a refactor.

On the RPi4, running one of my tunes with a full screen display on the
2560x1440 display, this reduces the CPU time taken for 1000 seconds worth of
play time by quite a bit:

 |          | master time | new time | diff | %          |
 |----------|-------------|----------|------|------------|
 | hydrogen | 453         | 368      | -85  | -18.8%     |
 | Xorg     | 83          | 15       | -68  | -81.9%     |
 | Total    | 536         | 383      | -153 | -28.5%     |

(So, overall that's going from 53% of a CPU core to 38% of a CPU core)