@theGreatWhiteShark 2022-08-25
Awesome!

I hope the performance didn't suffered too much due to the GUI update. After all, SVG rendering and direct drawing of components is most probably worse in terms of performance than loading PNG files. (I actively decided against checking the change in performance there in order to not discourage myself while updating the widgets :smile: )

We could introduce a checkable "Fullsize Playhead" option in the preferences to give better performance on old (or small) machines. If unchecked, the playhead will only the shown and updated in the ruler of both song and pattern editor. Indicating whether the current selected pattern is played back could also be represented in playhead color and would work there too.