@cme 2022-08-25
> Awesome!
> 
> I hope the performance didn't suffered too much due to the GUI update. After all, SVG rendering and direct drawing of components is most probably worse in terms of performance than loading PNG files. (I actively decided against checking the change in performance there in order to not discourage myself while updating the widgets :smile: )

It's probably worse, but I haven't compared it against anything before the check-in. Whatever the extra CPU overhead is, it's worth it! And theoretically if we find any real significant issues most of it can be changed to cached pixmaps if really necessary. I don't think it would be though. Avoiding unnecessary updates is proving to be fruitful in improving things. 

There's some more spurious redrawing in some circumstances I'm looking into but the largest single chunk of update time now seems to be drawing the LCD text elements, which isnt so trivial to do something about. 


> We could introduce a checkable "Fullsize Playhead" option in the preferences to give better performance on old (or small) machines. If unchecked, the playhead will only the shown and updated in the ruler of both song and pattern editor. Indicating whether the current selected pattern is played back could also be represented in playhead color and would work there too.

I think it should be fine even on older machines. The Pi4 is still really pretty slow in comparison to almost any hardware folks are likely to be using these days, and even then I had to push it to an edge case (full screen updating constantly) to get it to perform so badly as that.

What *might* be worth considering, though, is making the timers that drive the GUI updates adaptive so that they will update less frequently in slower hardware (and maybe a bit more smoothly on newer hardware).

Hmm... actually, yes, I like that, although it would mean I'd have to change my measurement method :D