@theGreatWhiteShark 2022-08-26
> Whatever the extra CPU overhead is, it's worth it!

That's true. I'm quite glad that there is no more pixel art involved in creating/changing a button.

> There's some more spurious redrawing in some circumstances I'm looking into but the largest single chunk of update time now seems to be drawing the LCD text elements, which isnt so trivial to do something about.

LCDs? Hmm. I would have expected something like the PPMs  in the mixer. The only LCD we constantly update - apart from peak levels in the mixer - should be the time display in the PlayerControl. Do you see a large drop in performance when opening the mixer? If not, this might indicate that LCD update is not properly checking for actual changes before repainting the widget.

There are also quite a number of timer-based updates in the GUI that could/should be rewrote to be triggered by core events. I almost started killing the timer in the instrument list of the pattern editor 2-3 times but decided to instead not delaying 1.2 further each time.

> What might be worth considering, though, is making the timers that drive the GUI updates adaptive so that they will update less frequently in slower hardware (and maybe a bit more smoothly on newer hardware).

Yeah. That sounds really nice!