@cme 2022-08-27
> LCDs? Hmm. I would have expected something like the PPMs in the mixer. The only LCD we constantly update - apart from peak levels in the mixer - should be the time display in the PlayerControl. Do you see a large drop in performance when opening the mixer? 

Aha, yes good point -- I did not have the mixer open. That adds a lot of time too. 

Having a quick look at the profile on macOS (where it's easier to profile with `sample` but the results are more variable than linux on the Pi4), it's dominated by the peak updates, which are dominated by the SVG rendering for the frame/background and the knob. We can fairly easily cache those to a single overlay pixmap (with an alpha channel) but not sure how much that would save unless I try it. 

> If not, this might indicate that LCD update is not properly checking for actual changes before repainting the widget.

It does check correctly, I already found that a while ago. :) I think it's mostly time-consuming because it's a large area of anti-aliased filled curve rendering; maybe there's some caching in the font rendering system that it's too large for. 

I'm going to try and get a profile build working on Linux to get a better idea. Apparently we *can* get profiling results from gprof for multithreaded processes, only we have to make sure the interval timer is set up for the thread. I'm not sure about thread safety either but we can profile one thread at a time to get round that.


> There are also quite a number of timer-based updates in the GUI that could/should be rewrote to be triggered by core events. I almost started killing the timer in the instrument list of the pattern editor 2-3 times but decided to instead not delaying 1.2 further each time.

On which note, yes, I'll do beta release process before I look at anything further ;)