@theGreatWhiteShark 2022-08-28
> I think it's mostly time-consuming because it's a large area of anti-aliased filled curve rendering; maybe there's some caching in the font rendering system that it's too large for.

In theory we could split the time display into individual widgets for ms, s, min etc. This would decrease the update area. But it would also make - at least IMO - the time display less appealing and probably goes too far in terms of optimization effort.

> On which note, yes, I'll do beta release process before I look at anything further ;)

Hehe. But there is definitely a lot of room to argue that these are important changes that deserve to be part of 1.2 and need to be tested properly via a beta.

In addition, I just came across some flaws in the current NSM implementation and I'm working on a patch that needs to be part of the upcoming beta (it introduces another translatable string/warning popup).