@cme 2022-08-28
> Hehe. But there is definitely a lot of room to argue that these are important changes that deserve to be part of 1.2 and need to be tested properly via a beta.
>
> In addition, I just came across some flaws in the current NSM implementation and I'm working on a patch that needs to be part of the upcoming beta (it introduces another translatable string/warning popup).

I'll hold off for now then! But also I've found some odd timing issues in current master, including transport locations being set to conflicting values from different places in audioengine.cpp (I found this by noticing multiple calls to update functions in profiling), and also song timing breaks somehow in songs with different length patterns (which I found when transcribing a complex time tune for actually playing tomorrow!)

So, I might need your help with those :D