@cme 2022-08-30
The weird behaviour on the complex time time tune doesn't seem to be deterministic, or perhaps it's related to changing pattern length somehow. I managed to reproduce it only one other time, when starting from scratch and adding empty patterns of the right length. (7,7,7,5,7,7,78)